track1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.3
betfairlightweight[speed]==2.21.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.21.1
